Added policies

+++ Task20/app/Http/Controllers/ServiceController.php
@@ -10,13 +10,10 @@ use Illuminate\Http\Request;
 
 class ServiceController extends Controller
 {
-    public function __construct()
-    {
-        $this->middleware(['admin']);
-    }
-
     public function index(Request $request)
     {
+        $this->authorize('delete', Service::class);
+
         $filter = new Filter();
 
         $services = $filter->run($request, Service::class)
@@ -28,6 +25,8 @@ class ServiceController extends Controller
 
     public function store(StoreServiceRequest $request)
     {
+        $this->authorize('create', Service::class);
+
         Service::create($request->validated());
 
         return redirect()->route('services');
@@ -35,11 +34,15 @@ class ServiceController extends Controller
 
     public function edit(Service $service)
     {
+        $this->authorize('edit', Service::class);
+
         return view('services.edit', ['service' => $service]);
     }
 
     public function update(UpdateServiceRequest $request, Service $service)
     {
+        $this->authorize('update', Service::class);
+
         $service->update($request->validated());
 
         return redirect()->route('services');
@@ -47,6 +50,8 @@ class ServiceController extends Controller
 
     public function destroy(Service $service)
     {
+        $this->authorize('delete', Service::class);
+
         $service->delete();
 
         return redirect()->route('services');
